# app/routes/user.py
from flask import Blueprint, request, jsonify
from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
from werkzeug.security import generate_password_hash, check_password_hash
from app import db
from app.models.user import User
from app.models.post import Post
from app.models.review import Review
from app.models.trade import Trade,TradeStatus
from datetime import datetime

user_bp = Blueprint('user', __name__, url_prefix='/api/v1/users')

# 회원가입
@user_bp.route('/register', methods=['POST'])
def register():
    data = request.get_json()
    username = data.get('username')
    email = data.get('email')
    password = data.get('password')

    if not all([username, email, password]):
        return jsonify({"msg": "필수 입력 항목이 누락되었습니다."}), 400

    if User.query.filter((User.username == username) | (User.email == email)).first():
        return jsonify({"msg": "이미 존재하는 사용자입니다."}), 409

    hashed_pw = generate_password_hash(password)
    user = User(username=username, email=email, password=hashed_pw)
    db.session.add(user)
    db.session.commit()

    return jsonify({"msg": "회원가입 완료"}), 201

# 로그인
@user_bp.route('/login', methods=['POST'])
def login():
    data = request.get_json()
    username = data.get('username')
    password = data.get('password')

    user = User.query.filter_by(username=username).first()
    if not user or not check_password_hash(user.password, password):
        return jsonify({"msg": "사용자 이름 또는 비밀번호가 올바르지 않습니다."}), 401

    access_token = create_access_token(identity=str(user.id))
    return jsonify({
        "access_token": access_token,
        "user": {
            "id": user.id,
            "username": user.username,
            "email": user.email
        }
    }), 200

# 로그아웃 (클라이언트에서 토큰 제거만으로 처리)
@user_bp.route('/logout', methods=['POST'])
@jwt_required()
def logout():
    user_id = get_jwt_identity()
    return jsonify({"msg": f"User {user_id} logged out (토큰 클라이언트에서 제거 필요)"}), 200

# 프로필 조회
@user_bp.route('/<int:userid>/profile', methods=['GET'])
def user_profile(userid):
    user = User.query.get(userid)
    if not user:
        return jsonify({"msg": "사용자를 찾을 수 없습니다."}), 404

    # 요청자 또는 수락자로 참여 중인 PENDING, ACCEPTED 상태의 거래 조회
    ongoing_trades = Trade.query.filter(
        ((Trade.requester_id == userid) | (Trade.receiver_id == userid)) &
        ((Trade.status == TradeStatus.PENDING) | (Trade.status == TradeStatus.ACCEPTED))
    ).all()

    # 거래 제목(게시글에서 가져옴)과 썸네일 이미지 준비
    ongoing_trades_data = []
    for trade in ongoing_trades:
        trade_data = {
            "trade_id": trade.id,
            "title": trade.post.title if trade.post else "제목 없음",
            "status": trade.status.value,
            "thumbnail_image_url": trade.post.thumbnail_image_url if trade.post else "default.jpg"
        }
        ongoing_trades_data.append(trade_data)

    return jsonify({
        "user": {
            "id": user.id,
            "username": user.username,
            "email": user.email,
            "profile_image_url": user.profile_image_url or "default.jpg",
            "registeredAt": user.created_at.strftime('%Y-%m-%d') if user.created_at else "",
            "current_points": user.points,
            "current_grade": user.grade or "보따리장수",
            "grade_icon_url": f"{user.grade.lower()}_grade.png" if user.grade else "보따리장수_grade.png",
            "total_trades": len(user.trades),
            "completed_trades": len([t for t in user.trades if t.status == 'COMPLETED']),
            "cancellation_count": 0,
            "cancellation_warning": "false",
            "ongoing_trades": ongoing_trades_data
        }
    })

# 리뷰 목록
@user_bp.route('/<int:userid>/reviews', methods=['GET'])
def user_review(userid):
    reviews = Review.query.filter_by(user_id=userid).order_by(Review.created_at.desc()).all()

    result = [{
        "review_id": r.id,
        "reviewer_id": r.reviewer_id,
        "rating": r.rating,
        "comment": r.comment,
        "date": r.created_at.strftime('%Y-%m-%d')
    } for r in reviews]

    return jsonify({
        "status": "success",
        "user_id": userid,
        "review_count": len(reviews), 
        "reviews": result
    }), 200

# 거래 이력
@user_bp.route('/<int:userid>/trades', methods=['GET'])
@jwt_required()
def user_trade_history(userid):
    """
    현재 로그인한 사용자가 참여한 거래 내역을 반환합니다.
    응답 예시:
    [
      {
        "tradeId": 123,
        "postId": 456,
        "postTitle": "중고 노트북 팝니다",
        "counterpartId": 789,
        "status": "COMPLETED",
        "createdAt": "2025-05-31 20:20",
        "completedAt": "2025-06-01 10:00"
      },
      ...
    ]
    """
    user = User.query.get(userid)
    if not user:
        return jsonify({"msg": "사용자를 찾을 수 없습니다."}), 404

    # 로그인한 사용자가 requester 또는 receiver인 모든 Trade를 조회
    trades = Trade.query.filter(
        (Trade.requester_id == userid) | (Trade.receiver_id == userid)
    ).order_by(Trade.created_at.desc()).all()

    result = []
    for t in trades:
        # 관련된 Post 객체
        post = t.post  # Trade 모델에서 post = relationship('Post', …) 로 연결되어 있어야 합니다
        if not post:
            # 혹시 post가 삭제되어 없을 수도 있으니, 안전하게 건너뜁니다
            continue

        # 거래 상대방 ID 계산
        counterpart_id = t.receiver_id if t.requester_id == userid else t.requester_id

        result.append({
            "tradeId": t.id,
            "postId": post.id,
            "postTitle": post.title,        # ← t.post.title 로 수정
            "counterpartId": counterpart_id,
            "status": t.status.value,       # TradeStatus enum이면 .value 를 사용
            "createdAt": t.created_at.strftime('%Y-%m-%d %H:%M'),
            "completedAt": t.completed_at.strftime('%Y-%m-%d %H:%M') if t.completed_at else None,
            "role": "buyer" if t.requester_id == userid else "seller"
        })

    return jsonify(result), 200

# 포인트 수정
@user_bp.route('/<int:userid>/points', methods=['PATCH'])
@jwt_required()
def user_points(userid):
    if int(get_jwt_identity()) != userid:
        return jsonify({"msg": "권한이 없습니다."}), 403

    data = request.get_json()
    amount = data.get("amount")

    if amount is None:
        return jsonify({"msg": "필드 누락"}), 400

    user = User.query.get(userid)
    if not user:
        return jsonify({"msg": "사용자 없음"}), 404

    user.points += amount
    db.session.commit()

    return jsonify({"msg": "포인트 수정 성공"}), 200

# 등급 조회
@user_bp.route('/<int:userid>/grade', methods=['GET'])
def user_grade(userid):
    user = User.query.get(userid)

    if not user:
        return jsonify({"msg": "사용자 없음"}), 404

    if user.points >= 3000:
        grade = "교환장인"
    elif user.points >= 2000:
        grade = "장사꾼"
    elif user.points >= 1000:
        grade = "초보상인"
    else:
        grade = "보따리장수"
    user.grade=grade
    db.session.commit()
    return jsonify({
        "points": user.points,
        "grade": user.grade,
        "gradeIconUrl": f"{grade}_grade.png"
    }), 200

# 회원 탈퇴
@user_bp.route('/<int:userid>', methods=['DELETE'])
@jwt_required()
def delete_user(userid):
    if int(get_jwt_identity()) != userid:
        return jsonify({"msg": "권한이 없습니다."}), 403

    user = User.query.get(userid)
    if not user:
        return jsonify({"msg": "사용자 없음"}), 404

    db.session.delete(user)
    db.session.commit()

    return jsonify({"msg": "회원 탈퇴 완료"}), 200

@user_bp.route('/<int:userid>/categories', methods=['PATCH'])
@jwt_required()
def update_categories(userid):
    if int(get_jwt_identity()) != userid:
        return jsonify({"msg": "권한이 없습니다."}), 403

    data = request.get_json()
    categories = data.get('categories')  # 리스트로 받음

    if not isinstance(categories, list) or len(categories) > 5:
        return jsonify({"msg": "카테고리는 리스트 형식이며 최대 5개까지 가능합니다."}), 400

    user = User.query.get(userid)
    if not user:
        return jsonify({"msg": "사용자를 찾을 수 없습니다."}), 404

    # 카테고리를 순서대로 할당
    for i in range(5):
        setattr(user, f'category{i+1}', categories[i] if i < len(categories) else None)

    db.session.commit()

    return jsonify({
        "msg": "카테고리 업데이트 완료",
        "categories": [user.category1, user.category2, user.category3, user.category4, user.category5]
    }), 200

@user_bp.route('/<int:userid>/categories', methods=['GET'])
@jwt_required()
def get_categories(userid):
    if int(get_jwt_identity()) != userid:
        return jsonify({"msg": "권한이 없습니다."}), 403

    user = User.query.get(userid)
    if not user:
        return jsonify({"msg": "사용자를 찾을 수 없습니다."}), 404

    categories = [
        user.category1,
        user.category2,
        user.category3,
        user.category4,
        user.category5
    ]

    return jsonify({
        "msg": "카테고리 조회 성공",
        "categories": categories
    }), 200

# 회원정보 수정
@user_bp.route('/<int:userid>/profile', methods=['PATCH'])
@jwt_required()
def update_user_profile(userid):
    if int(get_jwt_identity()) != userid:
        return jsonify({"msg": "권한이 없습니다."}), 403

    user = User.query.get(userid)
    if not user:
        return jsonify({"msg": "사용자를 찾을 수 없습니다."}), 404

    data = request.get_json()
    username = data.get("username")
    email = data.get("email")
    profile_image_url = data.get("profile_image_url")

    if username:
        if User.query.filter(User.username == username, User.id != userid).first():
            return jsonify({"msg": "이미 사용 중인 닉네임입니다."}), 409
        user.username = username

    if email:
        if User.query.filter(User.email == email, User.id != userid).first():
            return jsonify({"msg": "이미 사용 중인 이메일입니다."}), 409
        user.email = email

    if profile_image_url is not None:
        user.profile_image_url = profile_image_url

    db.session.commit()

    return jsonify({
        "msg": "회원정보 수정 완료",
        "user": {
            "id": user.id,
            "username": user.username,
            "email": user.email,
            "profile_image_url": user.profile_image_url
        }
    }), 200