# openapi.yaml
openapi: 3.0.0
info:
  title: TradeOn API
  version: 1.0.0
  description: 'TradeOn은 사용자 간 물물교환 기반의 중고 거래 플랫폼입니다.


    이 API 문서는 다음과 같은 기능들을 포함합니다:

    - 사용자 회원가입, 로그인 및 정보 조회

    - 교환 요청 및 거래 이력 관리

    - 사용자 리뷰 시스템

    - 가치 평가 게시판

    - 하이브리드 검색기능

    - 게시물 등록/삭제/조회

    - 채팅 및 거래 약속

    - 거래 데이터 분석 및 시각화


    JWT 인증이 필요한 API의 경우 Authorize 버튼을 통해 토큰을 입력해주세요.'
servers:
- url: https://673e-121-140-48-141.ngrok-free.app
  description: 로컬
...
paths:
  /api/v1/users/register:
    post:
      tags:
        - User
      summary: 사용자 회원가입 (id=userid)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: 회원가입 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: 회원가입 완료
        '400':
          description: 필수 입력 항목 누락
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 이미 존재하는 사용자
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/login:
    post:
      tags:
        - User
      summary: 사용자 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
        '401':
          description: 로그인 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/logout:
    post:
      tags:
        - User
      summary: 사용자 로그아웃 (ex.localStorage.removeItem('access_token') 등으로 클라이언트에서 토큰 제거해야함 )
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 로그아웃 처리 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: User logged out (토큰 클라이언트에서 제거 필요)
  /api/v1/users/{userid}/profile:
    get:
      tags:
        - User
      summary: 사용자 프로필 조회 (userid(int) 입력)
      parameters:
        - name: userid
          in: path
          description: 조회할 사용자 ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 프로필 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - User
      summary: 사용자 회원정보 수정
      description: 인증된 사용자가 자신의 닉네임, 이메일, 프로필 이미지를 수정합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          description: 수정할 사용자 ID
          required: true
          schema:
            type: integer
      requestBody:
        description: 수정할 사용자 정보(모든 필드 선택적)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        "200":
          description: 회원정보 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserProfileResponse'
        "403":
          description: 권한 없음 (다른 사용자의 프로필 수정 시도)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: 해당 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: 충돌 (이미 사용 중인 닉네임 또는 이메일)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'            
  /api/v1/users/{userid}/reviews:
    get:
      tags:
        - User
      summary: 사용자 리뷰 조회 (userid(int)입력)
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 리뷰 리스트 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserReviewListResponse'
  /api/v1/users/{userid}/trades:
    get:
      tags:
        - User
      summary: 사용자 거래 내역 조회 (userid(int) 입력)
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
          description: 조회할 사용자 ID
      responses:
        "200":
          description: 거래 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTrade'
        "401":
          description: 인증 필요 또는 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: 사용자가 없거나 거래 내역이 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userid}/points:
    patch:
      tags:
        - User
      summary: 사용자 포인트 수정 ('amount 값 1~50 제한')
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPointsUpdateRequest'
      responses:
        '200':
          description: 포인트 수정 성공
        '400':
          description: 필드 누락
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userid}/grade:
    get:
      tags:
        - User
      summary: 사용자 등급 조회
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 등급 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGradeResponse'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userid}:
    delete:
      tags:
        - User
      summary: 사용자 탈퇴
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 탈퇴 완료
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userid}/categories:
    patch:
      tags:
        - User
      summary: 사용자 관심 카테고리 업데이트
      description: 인증된 사용자가 자신의 카테고리 목록을 최대 5개까지 수정합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          description: 수정할 사용자 ID
          required: true
          schema:
            type: integer
      requestBody:
        description: 업데이트할 카테고리 목록 (최대 5개)
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        "200":
          description: 카테고리 업데이트 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryUpdateResponse'
        "400":
          description: 잘못된 요청 (카테고리 형식 오류 또는 개수 초과)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: 인증 필요 또는 토큰 만료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: 권한 없음 (다른 사용자의 카테고리 수정 시도)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: 해당 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - User
      summary: 사용자 관심 카테고리 조회
      description: 인증된 사용자가 자신의 관심 카테고리(최대 5개)를 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          description: 조회할 사용자 ID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 카테고리 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        "403":
          description: 권한 없음 (다른 사용자의 카테고리 열람 시도)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: 해당 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/analytics/user/{userid}:
    get:
      tags:
        - Analytic
      summary: 특정 사용자 거래 및 평점 통계 조회
      description: |
        지정된 userId를 가진 사용자의 총 거래 건수, 성공 거래 건수, 성공률, 평균 평점,
        월별 완료된 거래 건수 및 월별 평균 평점을 조회하고 결과를 DB에 저장하거나 업데이트합니다.
        JWT 인증이 필요합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          description: 통계를 조회할 사용자 ID
          schema:
            type: integer
            example: 42
      responses:
        "200":
          description: 통계 조회 및 업데이트 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        "401":
          description: 인증되지 않음 또는 토큰이 유효하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: 사용자 또는 거래 내역이 존재하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  
    put:
      tags:
        - Analytic
      summary: 특정 사용자 거래 데이터 수동 업데이트
      description: |
        지정된 userId를 가진 사용자의 총 거래 건수, 성공 거래 건수, 평균 평점을
        요청 바디에 포함된 값으로 수동 업데이트합니다. 성공률은 자동 계산되어 저장됩니다.
        JWT 인증이 필요합니다.-> 관리자용이므로 실제 구현할 필요없다고 판단
      security:
        - bearerAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          description: 업데이트할 사용자 ID
          schema:
            type: integer
            example: 42
      requestBody:
        description: 수동으로 업데이트할 거래 데이터 값
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAnalyticsRequest'
      responses:
        "200":
          description: 거래 데이터 수동 업데이트 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAnalyticsResponse'
        "400":
          description: 잘못된 요청 (필수 필드 누락 등)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: 인증되지 않음 또는 토큰이 유효하지 않음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    UserRegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: user123
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: password123
    UserLoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: user123
        password:
          type: string
          example: password123
    UserLoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          type: object
          properties:
            id:
              type: integer
              example: 1
            username:
              type: string
              example: user123
            email:
              type: string
              example: user@example.com
    UserProfileResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: user123
        email:
          type: string
          example: user@example.com
        image:
          type: string
          example: https://example.com/profile.png
        ongoing_trades:
              type: array
              description: 진행 중인 거래 목록
              items:
                type: object
                properties:
                  trade_id:
                    type: integer
                    description: 거래 ID
                    example: 42
                  title:
                    type: string
                    description: 거래 게시글 제목
                    example: "중고 노트북 팝니다"
                  status:
                    type: string
                    description: 거래 상태 (PENDING 또는 ACCEPTED)
                    example: "PENDING"
                  thumbnail_image_url:
                    type: string
                    description: 거래 게시글 썸네일 이미지 URL
                    example: https://example.com/thumb.jpg
                required:
                  - trade_id
                  - title
                  - status
                  - thumbnail_image_url
    UpdateUserProfileRequest:
      type: object
      properties:
        username:
          type: string
          description: 변경할 닉네임 (선택)
          example: "new_username"
        email:
          type: string
          description: 변경할 이메일 (선택)
          format: email
          example: "new_email@example.com"
        profile_image_url:
          type: string
          description: 변경할 프로필 이미지 URL (선택)
          example: "/static/uploads/new_profile.png"
      # 모두 선택적이므로 required 없음
    UpdateUserProfileResponse:
      type: object
      properties:
        msg:
          type: string
          description: 응답 메시지
          example: "회원정보 수정 완료"
        user:
          type: object
          properties:
            id:
              type: integer
              description: 사용자 ID
              example: 1
            username:
              type: string
              description: 수정된 닉네임
              example: "new_username"
            email:
              type: string
              description: 수정된 이메일
              example: "new_email@example.com"
            profile_image_url:
              type: string
              description: 수정된 프로필 이미지 URL
              example: "/static/uploads/new_profile.png"
          required:
            - id
            - username
            - email
            - profile_image_url
      required:
        - msg
        - user              
    UserReviewListResponse:
      type: object
      properties:
        status:
          type: string
          description: 요청 처리 상태
          example: "success"
        user_id:
          type: integer
          description: 조회한 사용자 ID
          example: 5
        review_count:
          type: integer
          description: 총 리뷰 개수
          example: 3
        reviews:
          type: array
          description: 리뷰 목록
          items:
            $ref: '#/components/schemas/UserReview'
      required:
        - status
        - user_id
        - review_count
        - reviews
    UserReview:
      type: object
      properties:
        review_id:
          type: integer
          description: 리뷰 ID
          example: 42
        reviewer_id:
          type: integer
          description: 리뷰어(작성자) 사용자 ID
          example: 7
        rating:
          type: integer
          description: 평점(예-> 1~5 사이)
          example: 4
        comment:
          type: string
          description: 리뷰 코멘트
          example: "정말 친절한 거래였습니다!"
        date:
          type: string
          format: date
          description: 리뷰 작성일 (YYYY-MM-DD)
          example: "2025-05-31"
      required:
        - review_id
        - reviewer_id
        - rating
        - comment
        - date
    UserTrade:
      type: object
      properties:
        tradeId:
          type: integer
          description: 거래 ID
          example: 123
        postId:
          type: integer
          description: 거래된 게시글 ID
          example: 456
        postTitle:
          type: string
          description: 거래된 게시글 제목
          example: "중고 노트북 팝니다"
        counterpartId:
          type: integer
          description: 거래 상대방 사용자 ID
          example: 789
        status:
          type: string
          description: 거래 상태 (PENDING, ACCEPTED, COMPLETED 중 하나)
          example: "COMPLETED"
        createdAt:
          type: string
          format: date-time
          description: 거래 생성 시각 (YYYY-MM-DD HH:MM)
          example: "2025-05-31 20:20"
        completedAt:
          type: string
          format: date-time
          description: 거래 완료 시각 (거래 완료 전에는 null)
          nullable: true
          example: "2025-06-01 10:00"
      required:
        - tradeId
        - postId
        - postTitle
        - counterpartId
        - status
        - createdAt
    UserPointsUpdateRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: integer
          example: 150
    UserGradeResponse:
      type: object
      properties:
        grade:
          type: string
          example: 초보상인
        points:
          type: integer
          example: 1250
    CategoryUpdateRequest:
      type: object
      properties:
        categories:
          type: array
          description: 최대 5개의 카테고리 문자열을 순서대로 전달
          items:
            type: string
          minItems: 0
          maxItems: 5
          example: ["운동", "요리", "음악"]
      required:
        - categories
    CategoryUpdateResponse:
      type: object
      properties:
        msg:
          type: string
          description: 성공 메시지
          example: "카테고리 업데이트 완료"
        categories:
          type: array
          description: 현재 저장된 5개의 카테고리 (값이 없으면 null)
          items:
            type: string
            nullable: true
          example: ["운동", "요리", "null", "null", "null"]
      required:
        - msg
        - categories    
    CategoryListResponse:
      type: object
      properties:
        msg:
          type: string
          description: 응답 메시지
          example: "카테고리 조회 성공"
        categories:
          type: array
          description: 사용자가 설정한 최대 5개의 카테고리 (값이 없으면 null)
          items:
            type: string
            nullable: true
          example: ["전자기기", "도서", null, null, null]
      required:
        - msg
        - categories    
...
    AnalyticsResponse:
      type: object
      properties:
        userId:
          type: integer
          description: 사용자 ID
          example: 42
        totalTrades:
          type: integer
          description: 총 거래 건수
          example: 10
        successfulTrades:
          type: integer
          description: 성공 거래(Completed) 건수
          example: 7
        successRate:
          type: number
          format: float
          description: 성공 거래 비율(퍼센트)
          example: 70.0
        averageRating:
          type: number
          format: float
          description: 해당 사용자가 받은 평균 평점
          example: 4.5
        monthlyTradeCounts:
          type: array
          description: 월별 완료된 거래 건수 리스트
          items:
            type: object
            properties:
              month:
                type: string
                format: date
                description: 연-월 (YYYY-MM 형식)
                example: "2025-05"
              count:
                type: integer
                description: 해당 월에 완료된 거래 수
                example: 3
        monthlyAverageRatings:
          type: array
          description: 월별 평균 평점 리스트
          items:
            type: object
            properties:
              month:
                type: string
                format: date
                description: 연-월 (YYYY-MM 형식)
                example: "2025-05"
              averageRating:
                type: number
                format: float
                description: 해당 월에 계산된 평균 평점
                example: 4.2
      required:
        - userId
        - totalTrades
        - successfulTrades
        - successRate
        - averageRating
        - monthlyTradeCounts
        - monthlyAverageRatings
    UpdateAnalyticsRequest:
      type: object
      properties:
        totalTrades:
          type: integer
          description: 업데이트할 총 거래 건수
          example: 10
        successfulTrades:
          type: integer
          description: 업데이트할 성공 거래 건수
          example: 7
        averageRating:
          type: number
          format: float
          description: 업데이트할 평균 평점
          example: 4.5
      required:
        - totalTrades
        - successfulTrades
        - averageRating
    UpdateAnalyticsResponse:
      type: object
      properties:
        msg:
          type: string
          description: 상태 메시지
          example: "거래 데이터 수동 업데이트 완료"
        userId:
          type: integer
          description: 업데이트된 사용자 ID
          example: 42
        totalTrades:
          type: integer
          description: 업데이트된 총 거래 건수
          example: 10
        successfulTrades:
          type: integer
          description: 업데이트된 성공 거래 건수
          example: 7
        averageRating:
          type: number
          format: float
          description: 업데이트된 평균 평점
          example: 4.5
      required:
        - msg
        - userId
        - totalTrades
        - successfulTrades
        - averageRating
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        msg:
          type: string
          example: 오류 메시지입니다.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []